+++ changelog.md
@@ -17,6 +17,8 @@ Codion Change Log
 - TemporalFieldPanel, TextFieldPanel.setTransferFocusOnEnter() renamed transferFocusOnEnter().
 ### is.codion.swing.framework.model
 - SwingEntityTableModel.EntityFilterModelFactory now creates String based column condition models for entity, item and non-comparable based columns.
+### is.codion.swing.framework.model
+- SwingEntityModel.Builder.modelClass(), editModelClass(), tableModelClass(), modelFactory(), editModelFactory(), tableModelFactory() and detailModelBuilder() renamed model(), editModel(), tableModel() and detailModel().
 ### is.codion.swing.framework.ui
 - TabbedApplicationLayout bug fixed, now applies the border to the application panel, instead of only the tabbed pane.
 - EntityTablePanel.DefaultStatusMessage bug fixed, no longer returns an empty message when all rows are filtered.
@@ -25,6 +27,7 @@ Codion Change Log
 - EntityDialogs.EditDialogBuilder.Updater now with generic edit model type parameter.
 - EntitySearchField.Selector.setPreferredSize() renamed preferredSize().
 - EntitySearchField.setSelectorFactory() renamed selectorFactory().
+- EntityPanel.Builder.panelClass(), editPanelClass(), tablePanelClass(), detailPanelBuilder() and panelLayout() renamed panel(), editPanel(), tablePanel(), detailPanel() and layout().
 
 ## 0.17.24
 ### is.codion.common.core